Revert "Conditionally include saved payment route,"
This reverts commit a4e5aaa01b63b0dbd43960c6e9309129a838363c.
