+++ packages/peregrine/lib/talons/AccountMenu/useAccountMenuItems.js
@@ -14,18 +14,7 @@ export const useAccountMenuItems = props => {
         onSignOut();
     }, [onSignOut]);
 
-    const eeOnlyMenuItems = [
-        {
-            name: 'Saved Payments',
-            id: 'accountMenu.savedPaymentsLink',
-            url: '/saved-payments'
-        }
-    ];
-
     const MENU_ITEMS = [
-        ...(process.env.MAGENTO_BACKEND_EDITION === 'EE'
-            ? eeOnlyMenuItems
-            : []),
         {
             name: 'Order History',
             id: 'accountMenu.orderHistoryLink',
@@ -45,6 +34,11 @@ export const useAccountMenuItems = props => {
             name: 'Address Book',
             id: 'accountMenu.addressBookLink',
             url: '/address-book'
+        },
+        {
+            name: 'Saved Payments',
+            id: 'accountMenu.savedPaymentsLink',
+            url: '/saved-payments'
         },
         {
             name: 'Communications',
