+++ packages/venia-ui/lib/targets/venia-ui-intercept.js
@@ -10,7 +10,6 @@ const RichContentRendererList = require('./RichContentRendererList');
  */
 
 const name = '@magento/venia-ui';
-const isEE = process.env.MAGENTO_BACKEND_EDITION === 'EE';
 
 module.exports = targets => {
     const builtins = targets.of('@magento/pwa-buildpack');
@@ -68,19 +67,9 @@ module.exports = targets => {
         });
     });
 
-    const eeOnlyRoutes = [
-        {
-            name: 'SavedPayments',
-            pattern: '/saved-payments',
-            exact: true,
-            path: '../SavedPaymentsPage'
-        }
-    ];
-
     // The paths below are relative to packages/venia-ui/lib/components/Routes/routes.js.
     targets.own.routes.tap(routes => [
         ...routes,
-        ...(isEE ? eeOnlyRoutes : []),
         {
             name: 'AddressBook',
             pattern: '/address-book',
@@ -126,6 +115,12 @@ module.exports = targets => {
             pattern: '/customer/account/createPassword',
             exact: true,
             path: '../MyAccount/ResetPassword'
+        },
+        {
+            name: 'SavedPayments',
+            pattern: '/saved-payments',
+            exact: true,
+            path: '../SavedPaymentsPage'
         },
         {
             name: 'Search',
